refactor: remove useless initing app function
