+++ src/redux/reducers.js
@@ -3,7 +3,7 @@ import { combineReducers } from 'redux';
 import robotReducer from '../containers/RobotPage/robotReducer';
 import { getAllChatContentReducer, relatedCurrentChatReducer } from './reducers/chatContentReducer';
 import { getHomePageListReducer } from '../containers/HomePageList/homePageListReducer';
-import { initAppReducer } from '../utils/InitApp/initAppReducer';
+import { initAppReducer } from './reducers/initAppReducer';
 
 export default combineReducers({
   robotState: robotReducer,
